+++ src/main/kotlin/cloud/glitchdev/rfu/data/hud/HudConfig.kt
@@ -1,6 +1,7 @@
 package cloud.glitchdev.rfu.data.hud
 
 class HudConfig {
+    var hasInitializedDefaults: Boolean = false
     var hudElements : MutableList<HudElement> = mutableListOf()
 
     fun getOrAdd(id : String, defaultX : Float = 0f, defaultY: Float = 0f, scale: Float = 1f) : HudElement {