+++ src/main/kotlin/cloud/glitchdev/rfu/feature/settings/Hud.kt
@@ -1,17 +1,25 @@
 package cloud.glitchdev.rfu.feature.settings
 
 import cloud.glitchdev.rfu.gui.window.HudWindow
+import cloud.glitchdev.rfu.utils.command.AbstractCommand
 import cloud.glitchdev.rfu.utils.command.Command
-import cloud.glitchdev.rfu.utils.command.SimpleCommand
-import com.mojang.brigadier.context.CommandContext
+import com.mojang.brigadier.builder.LiteralArgumentBuilder
 import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource
 
 @Command
-object Hud : SimpleCommand("rfumove") {
-    override val description: String = "Opens the GUI to move Hud Elements."
+object Hud : AbstractCommand("rfumove") {
+    override val description: String = "Opens the GUI to move Hud Elements or reset to defaults."
 
-    override fun execute(context: CommandContext<FabricClientCommandSource>): Int {
-        HudWindow.openEditingGui()
-        return 1
+    override fun build(builder: LiteralArgumentBuilder<FabricClientCommandSource>) {
+        builder.executes {
+            HudWindow.openEditingGui()
+            1
+        }
+        builder.then(
+            lit("reset").executes {
+                HudWindow.resetAllToDefaults()
+                1
+            }
+        )
     }
 }