+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added Party Finder party presets system
 - Added {mob} variable to rare drop chat messages and title alerts
 - Added Rare Drops Display
+- Added a better default HUD system and a reset button
 
 ### Fixes
 - Added Elementa memory leak patch (toggleable in Other settings if it causes conflicts)